+++ emborg/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.14.2'
-__released__ = '2020-01-02'
+__version__ = '1.14.3'
+__released__ = '2020-01-22'
 
 from .settings import Settings as Emborg